# contracts/http/openapi.yaml
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "EcoBin P0 HTTP API",
    "version": "1.0.0-rc.1",
    "description": "EcoBin /api/v1 transport contract. Business slices add their own paths and schemas without redefining these cross-cutting rules."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin deployment"
    }
  ],
...
  "paths": {
...
    "/api/v1/web/platform/tenants/{tenantCode}": {
      "parameters": [{"$ref": "#/components/parameters/TenantCode"}],
      "get": {
        "operationId": "getIdentityTenant",
        "tags": ["Identity directory"],
        "summary": "Read one tenant through the platform privilege boundary",
        "security": [{"webSessionCookie": []}],
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"}
        }
      }
    },
...
    "/api/v1/web/platform/tenants/{tenantCode}/organizations/{organizationCode}": {
      "parameters": [
        {"$ref": "#/components/parameters/TenantCode"},
        {"$ref": "#/components/parameters/OrganizationCode"}
      ],
      "get": {
        "operationId": "getPlatformIdentityOrganization",
        "tags": ["Identity directory"],
        "summary": "Read an organization in one explicit platform target tenant",
        "security": [{"webSessionCookie": []}],
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"}
        }
      }
    },
...
    "/api/v1/web/organizations/{organizationCode}/staff-accounts/{staffUid}/miniapp-binding": {
      "get": {
        "operationId": "getCurrentTenantStaffMiniappBinding",
        "tags": ["Identity directory"],
        "summary": "Read the staff account current miniapp binding snapshot",
        "security": [{"webSessionCookie": []}],
        "parameters": [
          {"$ref": "#/components/parameters/OrganizationCode"},
          {"$ref": "#/components/parameters/StaffAccountUid"}
        ],
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "403": {"$ref": "#/components/responses/ForbiddenProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"}
        }
      },
      "put": {
        "operationId": "setCurrentTenantStaffMiniappBinding",
        "tags": ["Identity directory"],
        "summary": "Atomically set or replace a staff miniapp binding using both observed snapshots",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "parameters": [
          {"$ref": "#/components/parameters/OrganizationCode"},
          {"$ref": "#/components/parameters/StaffAccountUid"},
          {"$ref": "#/components/parameters/IdempotencyKey"}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/SetStaffMiniappBindingRequest"}
            }
          }
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "403": {"$ref": "#/components/responses/ForbiddenProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"},
          "422": {"$ref": "#/components/responses/BusinessRuleProblem"}
        }
      }
    },
...
    "/api/v1/web/platform/tenants/{tenantCode}/organizations/{organizationCode}/organization-users/phone-lookups": {
      "post": {
        "operationId": "lookupPlatformOrganizationUserByPhone",
        "tags": ["Identity directory"],
        "summary": "Platform lookup of one organization user by a body-only phone number",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "parameters": [
          {"$ref": "#/components/parameters/TenantCode"},
          {"$ref": "#/components/parameters/OrganizationCode"}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/OrganizationUserPhoneLookupRequest"}
            }
          }
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"}
        }
      }
    },
    "/api/v1/web/platform/tenants/{tenantCode}/organizations/{organizationCode}/staff-accounts/{staffUid}/miniapp-binding": {
      "get": {
        "operationId": "getPlatformStaffMiniappBinding",
        "tags": ["Identity directory"],
        "summary": "Platform read of one staff miniapp binding snapshot",
        "security": [{"webSessionCookie": []}],
        "parameters": [
          {"$ref": "#/components/parameters/TenantCode"},
          {"$ref": "#/components/parameters/OrganizationCode"},
          {"$ref": "#/components/parameters/StaffAccountUid"}
        ],
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"}
        }
      },
      "put": {
        "operationId": "setPlatformStaffMiniappBinding",
        "tags": ["Identity directory"],
        "summary": "Platform atomically sets or replaces a staff miniapp binding",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "parameters": [
          {"$ref": "#/components/parameters/TenantCode"},
          {"$ref": "#/components/parameters/OrganizationCode"},
          {"$ref": "#/components/parameters/StaffAccountUid"},
          {"$ref": "#/components/parameters/IdempotencyKey"}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/SetStaffMiniappBindingRequest"}
            }
          }
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"},
          "422": {"$ref": "#/components/responses/BusinessRuleProblem"}
        }
      }
    },
    "/api/v1/web/platform/tenants/{tenantCode}/organizations/{organizationCode}/staff-miniapp-bindings/{bindingUid}/revocations": {
      "post": {
        "operationId": "revokePlatformStaffMiniappBinding",
        "tags": ["Identity directory"],
        "summary": "Platform revocation of one staff miniapp binding and its management sessions",
        "security": [{"webSessionCookie": [], "webCsrfToken": []}],
        "parameters": [
          {"$ref": "#/components/parameters/TenantCode"},
          {"$ref": "#/components/parameters/OrganizationCode"},
          {"$ref": "#/components/parameters/BindingUid"},
          {"$ref": "#/components/parameters/IdempotencyKey"}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/VersionCommand"}
            }
          }
        },
        "responses": {
          "200": {"$ref": "#/components/responses/DirectoryOk"},
          "400": {"$ref": "#/components/responses/InvalidRequest"},
          "401": {"$ref": "#/components/responses/UnauthorizedProblem"},
          "404": {"$ref": "#/components/responses/NotFoundProblem"},
          "409": {"$ref": "#/components/responses/ConflictProblem"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "webSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-ecobin-web-session",
        "description": "Secure, HttpOnly, SameSite=Lax, Path=/ and no Domain. JavaScript never reads this value."
      },
      "webCsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-TOKEN",
        "description": "Bootstrapped from GET /api/v1/web/auth/csrf-token and rotated after login/logout."
      },
      "miniappBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Exactly one short-lived aud=miniapp or aud=miniapp-staff session. No refresh token."
      },
      "wechatPaySignature": {
        "type": "apiKey",
        "in": "header",
        "name": "Wechatpay-Signature"
      },
      "wechatPayTimestamp": {
        "type": "apiKey",
        "in": "header",
        "name": "Wechatpay-Timestamp"
      },
      "wechatPayNonce": {
        "type": "apiKey",
        "in": "header",
        "name": "Wechatpay-Nonce"
      },
      "wechatPaySerial": {
        "type": "apiKey",
        "in": "header",
        "name": "Wechatpay-Serial"
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "UUIDv4 generated once for one human intent and reused by every retry of that same intent.",
        "schema": {
          "$ref": "#/components/schemas/UuidV4"
        }
      },
      "TenantCode": {
        "name": "tenantCode",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/TenantCode"
        }
      },
      "OrganizationCode": {
        "name": "organizationCode",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/OrganizationCode"
        }
      },
      "StaffAccountUid": {
        "name": "staffUid",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/PublicUid"
        }
      },
      "BindingUid": {
        "name": "bindingUid",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/PublicUid"
        }
      },
...
    },
    "headers": {
      "NoStore": {
        "description": "Sensitive or mutable projections are not cacheable",
        "schema": {
          "type": "string",
          "const": "no-store"
        }
      },
      "RequestId": {
        "description": "Transport request identifier; never a business idempotency identity",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128
        }
      },
...
    },
    "responses": {
      "DirectoryOk": {
        "description": "V-01 identity directory projection returned from current database facts",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          },
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DirectoryEnvelope"
            }
          }
        }
      },
...
      "InvalidRequest": {
        "description": "Malformed JSON, type, format, or cross-field input",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
      "UnauthorizedProblem": {
        "description": "Authentication is missing, invalid, revoked, expired, or has the wrong audience",
        "headers": {
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
      "ForbiddenProblem": {
        "description": "The visible resource exists but the authenticated subject lacks the current capability",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
      "NotFoundProblem": {
        "description": "The resource does not exist or is outside the caller's current visible scope",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
...
      "ConflictProblem": {
        "description": "Idempotency digest, expected version, state, or occupancy conflict",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
      "BusinessRuleProblem": {
        "description": "The input is valid but a current business qualification or invariant rejects it",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      },
...
    },
    "schemas": {
      "TenantCode": {
        "type": "string",
        "minLength": 1,
        "maxLength": 32,
        "pattern": "^[a-z0-9][a-z0-9-]*$"
      },
      "OrganizationCode": {
        "type": "string",
        "minLength": 1,
        "maxLength": 32,
        "pattern": "^[a-z0-9][a-z0-9-]*$"
      },
...
      "IdentityTenant": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "tenantCode",
          "enterpriseName",
          "status",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "tenantCode": {
            "$ref": "#/components/schemas/TenantCode"
          },
          "enterpriseName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "status": {
            "$ref": "#/components/schemas/DirectoryStatus"
          },
          "contactName": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 100
          },
          "contactPhone": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 32
          },
          "contactAddress": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 500
          },
          "version": {
            "$ref": "#/components/schemas/ExpectedVersion"
          },
          "principalAccount": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PrincipalAccountSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "$ref": "#/components/schemas/UtcTimestamp"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/UtcTimestamp"
          }
        }
      },
      "IdentityOrganization": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "organizationCode",
          "organizationName",
          "status",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "organizationCode": {
            "$ref": "#/components/schemas/OrganizationCode"
          },
          "organizationName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "status": {
            "$ref": "#/components/schemas/DirectoryStatus"
          },
          "contactPhone": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 32
          },
          "contactAddress": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 500
          },
          "version": {
            "$ref": "#/components/schemas/ExpectedVersion"
          },
          "createdAt": {
            "$ref": "#/components/schemas/UtcTimestamp"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/UtcTimestamp"
          }
        }
      },
      "StaffAccount": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "staffAccountUid",
          "accountKind",
          "loginName",
          "displayName",
          "status",
          "version",
          "authVersion",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "staffAccountUid": {
            "$ref": "#/components/schemas/PublicUid"
          },
          "accountKind": {
            "type": "string",
            "enum": [
              "TENANT_PRINCIPAL",
              "STAFF"
            ]
          },
          "loginName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "contactPhone": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 32
          },
          "status": {
            "$ref": "#/components/schemas/DirectoryStatus"
          },
          "version": {
            "$ref": "#/components/schemas/ExpectedVersion"
          },
          "authVersion": {
            "$ref": "#/components/schemas/ExpectedVersion"
          },
          "createdAt": {
            "$ref": "#/components/schemas/UtcTimestamp"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/UtcTimestamp"
          }
        }
      },
      "OrganizationMembership": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "organizationCode",
          "staffAccountUid",
          "displayName",
          "manager",
          "status",
          "permissionCodes",
          "version",
          "authVersion",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "organizationCode": {
            "$ref": "#/components/schemas/OrganizationCode"
          },
          "staffAccountUid": {
            "$ref": "#/components/schemas/PublicUid"
          },
          "displayName": {
            "type": "string"
          },
          "manager": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/DirectoryStatus"
          },
          "permissionCodes": {
            "$ref": "#/components/schemas/PermissionCodes"
          },
          "version": {
            "$ref": "#/components/schemas/ExpectedVersion"
          },
          "authVersion": {
            "$ref": "#/components/schemas/ExpectedVersion"
          },
          "createdAt": {
            "$ref": "#/components/schemas/UtcTimestamp"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/UtcTimestamp"
          }
        }
      },
...
      "EffectiveAccess": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "staffAccountUid",
          "tenantPermissionCodes",
          "organizations",
          "authVersion"
        ],
        "properties": {
          "staffAccountUid": {
            "$ref": "#/components/schemas/PublicUid"
          },
          "tenantPermissionCodes": {
            "$ref": "#/components/schemas/PermissionCodes"
          },
          "organizations": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "organizationCode",
                "organizationName",
                "manager",
                "permissionCodes"
              ],
              "properties": {
                "organizationCode": {
                  "$ref": "#/components/schemas/OrganizationCode"
                },
                "organizationName": {
                  "type": "string"
                },
                "manager": {
                  "type": "boolean"
                },
                "permissionCodes": {
                  "$ref": "#/components/schemas/PermissionCodes"
                }
              }
            }
          },
          "authVersion": {
            "$ref": "#/components/schemas/ExpectedVersion"
          }
        }
      },
...
      "ProvisionedOrganizationStaff": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "staffAccount",
          "membership"
        ],
        "properties": {
          "staffAccount": {
            "$ref": "#/components/schemas/StaffAccount"
          },
          "membership": {
            "$ref": "#/components/schemas/OrganizationMembership"
          }
        }
      },
      "OrganizationUserPhoneLookupRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["phoneNumber"],
        "properties": {
          "phoneNumber": {
            "type": "string",
            "minLength": 1,
            "maxLength": 24,
            "writeOnly": true,
            "description": "Exact phone number supplied only in the request body; never put it in a URL or log."
          }
        }
      },
      "BindingSnapshot": {
        "type": "object",
        "additionalProperties": false,
        "required": ["bindingUid", "version"],
        "properties": {
          "bindingUid": {"$ref": "#/components/schemas/PublicUid"},
          "version": {"type": "integer", "format": "int64", "minimum": 0}
        }
      },
...
      "OrganizationUserLookup": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "organizationUserUid",
          "nickname",
          "maskedPhoneNumber",
          "registeredAt",
          "status",
          "currentMiniappBinding"
        ],
        "properties": {
          "organizationUserUid": {"$ref": "#/components/schemas/PublicUid"},
          "nickname": {"type": "string", "minLength": 1, "maxLength": 100},
          "maskedPhoneNumber": {"type": "string", "minLength": 3, "maxLength": 24},
          "registeredAt": {"$ref": "#/components/schemas/UtcTimestamp"},
          "status": {"type": "string", "enum": ["ACTIVE", "FROZEN"]},
          "currentMiniappBinding": {
            "oneOf": [
              {"$ref": "#/components/schemas/OrganizationUserCurrentMiniappBinding"},
              {"type": "null"}
            ]
          }
        }
      },
...
      "OrganizationUser": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "organizationUserUid",
          "nickname",
          "avatarUrl",
          "maskedPhoneNumber",
          "phoneBound",
          "registeredAt",
          "registrationSource",
          "status",
          "cleanOperationEnabled",
          "version",
          "authVersion"
        ],
        "properties": {
          "organizationUserUid": {
            "$ref": "#/components/schemas/PublicUid"
          },
          "nickname": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "avatarUrl": {
            "oneOf": [
              {"type": "string", "format": "uri", "maxLength": 1000},
              {"type": "null"}
            ]
          },
          "maskedPhoneNumber": {
            "oneOf": [
              {"type": "string", "minLength": 3, "maxLength": 24},
              {"type": "null"}
            ]
          },
          "phoneBound": {"type": "boolean"},
          "registeredAt": {
            "$ref": "#/components/schemas/UtcTimestamp"
          },
          "registrationSource": {
            "oneOf": [
              {"$ref": "#/components/schemas/OrganizationUserRegistrationSource"},
              {"type": "null"}
            ]
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "FROZEN"]
          },
          "cleanOperationEnabled": {"type": "boolean"},
          "version": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "authVersion": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "StaffMiniappBindingLookup": {
        "type": "object",
        "additionalProperties": false,
        "required": ["currentMiniappBinding"],
        "properties": {
          "currentMiniappBinding": {
            "oneOf": [
              {"$ref": "#/components/schemas/StaffCurrentMiniappBinding"},
              {"type": "null"}
            ]
          }
        }
      },
      "SetStaffMiniappBindingRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "organizationUserUid",
          "expectedStaffBinding",
          "expectedOrganizationUserBinding"
        ],
        "properties": {
          "organizationUserUid": {"$ref": "#/components/schemas/PublicUid"},
          "expectedStaffBinding": {
            "oneOf": [
              {"$ref": "#/components/schemas/BindingSnapshot"},
              {"type": "null"}
            ]
          },
          "expectedOrganizationUserBinding": {
            "oneOf": [
              {"$ref": "#/components/schemas/BindingSnapshot"},
              {"type": "null"}
            ]
          },
          "reason": {
            "oneOf": [
              {"type": "string", "maxLength": 500},
              {"type": "null"}
            ]
          }
        }
      },
      "StaffMiniappBinding": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "bindingUid",
          "organizationUserUid",
          "staffAccountUid",
          "status",
          "version",
          "boundAt",
          "revokedAt"
        ],
        "properties": {
          "bindingUid": {"$ref": "#/components/schemas/PublicUid"},
          "organizationUserUid": {"$ref": "#/components/schemas/PublicUid"},
          "staffAccountUid": {"$ref": "#/components/schemas/PublicUid"},
          "status": {"type": "string", "enum": ["ACTIVE", "REVOKED"]},
          "version": {"type": "integer", "format": "int64", "minimum": 0},
          "boundAt": {"$ref": "#/components/schemas/UtcTimestamp"},
          "revokedAt": {
            "oneOf": [
              {"$ref": "#/components/schemas/UtcTimestamp"},
              {"type": "null"}
            ]
          }
        }
      },
      "DirectoryEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string",
            "const": "OK"
          },
          "data": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/IdentityTenant"
              },
              {
                "$ref": "#/components/schemas/IdentityOrganization"
              },
              {
                "$ref": "#/components/schemas/StaffAccount"
              },
              {
                "$ref": "#/components/schemas/OrganizationMembership"
              },
              {
                "$ref": "#/components/schemas/EffectiveAccess"
              },
              {
                "$ref": "#/components/schemas/ProvisionedOrganizationStaff"
              },
              {
                "$ref": "#/components/schemas/OrganizationUserLookup"
              },
              {
                "$ref": "#/components/schemas/OrganizationUser"
              },
              {
                "$ref": "#/components/schemas/StaffMiniappBindingLookup"
              },
              {
                "$ref": "#/components/schemas/StaffMiniappBinding"
              },
              {
                "$ref": "#/components/schemas/DirectoryPage"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PermissionDefinition"
                }
              }
            ]
          },
          "requestId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        }
      },
...
      "VersionCommand": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "expectedVersion"
        ],
        "properties": {
          "expectedVersion": {
            "$ref": "#/components/schemas/ExpectedVersion"
          },
          "reason": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
...
      "UuidV4": {
        "type": "string",
        "format": "uuid",
        "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
        "example": "4b9e8c22-7d38-4e4b-a1b0-7a0e1656e001"
      },
      "PublicUid": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UuidV4"
          }
        ],
        "description": "Stable public identity; never an internal BIGINT primary key"
      },
...
      "ExpectedVersion": {
        "type": "integer",
        "minimum": 0,
        "description": "Last server version explicitly observed by the operator",
        "example": 3
      },
...
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "requestId",
          "retryable",
          "details"
        ],
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]*(\\.[A-Z][A-Z0-9_]*)+$",
            "example": "COMMON.IDEMPOTENCY_KEY_CONFLICT"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500,
            "example": "相同操作标识已绑定到不同请求"
          },
          "requestId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "example": "01J3EXAMPLE00000000000001"
          },
          "retryable": {
            "type": "boolean",
            "example": false
          },
          "details": {
            "type": "object",
            "additionalProperties": true,
            "example": {}
          }
        }
      },
...
    }
  },
...
}